+++ .gitignore
@@ -81,6 +81,7 @@ typings/
 # Nuxt.js build / generate output
 .nuxt
 dist
+out.html
 
 # Gatsby files
 .cache/
