+++ .npmignore
@@ -1,6 +1,5 @@
-src/
 tests/
-build/
+examples/
 .github/
 .eslintrc.js
 .eslintignore
