+++ examples/main/data/get-another-number.js
@@ -1,3 +1,3 @@
 module.exports = {
-  abellNumber: 25
+  numberFromJS: 35
 };
