+++ examples/minimal/in.abell
@@ -1,0 +1,2 @@
+{{ }}
+<html></html>
