--- src/example/templates/inject_example.abell
@@ -1,46 +1,0 @@
-{{ 
-  const { globalMeta, numberFromJS } = require('./siteData.js');
-  const { numberFromJSON } = require('./meta.json');
-  const path = require('path');
-  const siteType = "blog";
-  const a = 3;
-}}
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <title>{{ globalMeta.siteName }}</title>
-</head>
-<body>
-{{ 
-  const b = 5;
-  console.log(a + b);
-}}
-  This website is of type '{{
-    (() => {
-      if(siteType === 'blog') {
-        return "Personal Blog"
-      }else{
-        return "Unknown"
-      }
-    })();
-  }}'
-  <header>
-    when {{numberFromJS}} is added to {{numberFromJSON}}, we get {{numberFromJS + numberFromJSON}}
-    <h1>{{globalMeta.siteName}}</h1>
-    <div class="header-bio">
-      Abell Minima is a starter blog for AbellJS. <br/>This template is designed by <b>{{ globalMeta.name }}</b> you can follow him on <a href="https://twitter.com/{{ globalMeta.twitter }}">Twitter @{{ globalMeta.twitter }}</a>.
-    </div>
-  </header>
-  <main>
-    {{ console.log(numberFromJS) }}
-    <div class="article-container">
-      {{
-        $contentList.map(content => 
-          `<div class="article-name">${content.name}</div>`
-        ).join('\n')
-      }}
-    </div>
-  </main>
-</body>
-</html>
