--- src/example/templates/loop_example/components/nice.abell
@@ -1,1 +1,0 @@
-<div>{{ 'Hello' }}</div>
