--- src/example/templates/loop_example/components/random.js
@@ -1,2 +1,0 @@
-const foo = 'bar';
-module.exports = foo;
