--- src/example/templates/siteData.js
@@ -1,8 +1,0 @@
-module.exports = {
-  globalMeta: {
-    siteName: 'This is my siteName',
-    name: 'This is my name',
-    twitter: 'saurabhcodes'
-  },
-  numberFromJS: 35
-};
